pyproject.toml
[project]
name = "journal"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "nellaG", email = "seirios0107@gmail.com" }
]
dependencies = [
    "django>=5.1",
    "pre-commit>=3.8.0",
    "ruff>=0.6.1",
    "psycopg>=3.2.1",
    "djangorestframework>=3.15.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/journal"]
